Adding in a rough sketch of requestor Dockerfile
Unable to get this to run the hash-cracker demo, it's unable to find the yagna daemon connect via socket, but it works with golemsp status.  Needs further debugging but wanted make a checkpoint this far. Will update file as it becomes more functional, a lot of it is taken from a golem hackathon project of provider in a container, but the clients are similar enough it should still work.