Blender.py that can accept args
I modified the blender demo to accept a packet of arguments. Not exactly sure where this goes so putting it into main, feel free to move it to a more appropriate space. Example command to run this with args for the demo is 

python3 blender.py --name '{"scene_file": "/golem/resource/scene.blend","resolution1": "400","resolution2" : "200","use_compositing": "False","samples": "100","output_format": "PNG","RESOURCES_DIR": "/golem/resources","WORK_DIR": "/golem/work","OUTPUT_DIR": "/golem/output","outfilebasename" : "out","borderx1" : "0.0","borderx2" : "1.0","bordery1" : "0.0","bordery2" : "1.0"}'

I did have to kill the old parser to get this working, so logfiles can't have custom names, and change the hardcoded network of "rinkeby" and driver of "zksync", it may be worth revising that but wasn't able to find the old parser to incorporate it into the new one. Error checking should also be added in the future but for now this just has a lot of hard type conversions so the chance of things breaking is a little bit lower.

Main arg is a json object that maps all the needed values to keys which are used to set args within the script